changed vector of pointers to regular vector, so that the transforms can work properly
Returning vectors of pointers for columns and rows didn't allow to save the `old` values temporarily, so the transforms were "shadowing" the colors.

There is still something wrong, I think it has to do with the orientation of the faces. There might be an inconsistency with what I call rows, or columns for each face. Needs more debugging